backend/pop_pipeline/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
pytrends==4.9.2
praw==7.7.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2
httpx==0.27.0
python-dotenv==1.0.1
schedule==1.2.1
rapidfuzz==3.9.1
numpy==1.26.4
pandas==2.2.2
aiofiles==23.2.1
tenacity==8.3.0
